requirements.txt
sentry-sdk==2.19.2
python-dotenv==1.0.1